requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
openai
flask-cors==3.0.10
gunicorn==20.1.0
nltk==3.8.1
tiktoken==0.5.2
python-magic==0.4.24
requests==2.31.0
